requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
sqlalchemy==1.4.46
psycopg2-binary==2.9.6
argon2-cffi==21.3.0
python-jose[cryptography]==3.3.0
httpx==0.24.1